Moved out all functions except submit
